[NFC] Split Make-Style Dependency Emission from FrontendTool
